Incremement is completly broken
When passing an increment in the expression the * is passed first, completly ignoring the /. Update the code to check for the * and / and pass parsing to the increment function. The increment function also adds the icnrement number as a string instead of a number. so strings concatanate instead of add.